requirements.txt
# Core dependencies
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Document processing
python-docx>=0.8.11
PyPDF2>=3.0.0
openpyxl>=3.1.0

# Image processing and OCR
Pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
easyocr>=1.7.0

# Speech recognition
whisper>=1.1.0
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0

# Language processing
transformers>=4.30.0
tokenizers>=0.13.0
sentencepiece>=0.1.99
spacy>=3.6.0

# Local LLM
llama-cpp-python>=0.2.0
ctransformers>=0.2.0

# Data handling
pydantic>=2.0.0
python-multipart>=0.0.6

# Utilities
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 